Delete root dto package and add dto package -> model/something/dto/..
