--- src/main/java/com/server/EZY/dto/DtoTest.java
@@ -1,4 +1,0 @@
-package com.server.EZY.dto;
-
-public class DtoTest {
-}
